or1k: change stack variable naming in libgloss
Before a pointer to the top of the stack was required by sbrk to
determine whether sufficient space is available. Instead we now use
three variables in crt0.S:

 * stack_top: points to top of the stack

 * stack_bottom: points to bottom of the reserved stack

 * stack_size: The size of the stack, set by the define STACK_SIZE

This has the advantage, that we can simply compare to the stack_bottom
can later easily allocate multiple stacks for the different cores and
just point to top and bottom of the reserved area.