or1k: Use extra stack for exception handling
Use an extra stack for the exception handling. Reason: We cannot
simply use the current stack pointer. We therefore reserve an extra
system exception stack, that is only used during exceptions.

The loading of the exception stack address needs an extra (clean)
register. Discussion goes on on how to solve this issue. For the
moment we use address 0x4 in the global memory.