or1k: Dynamic reentrancy and exception impure data
With the previous implementation there was only one impure data
structure. Problems occur if an exception happens during a library
function. The impure data is then shared and may be corrupted.

This changes this behavior by implementing two different impure data
structures for the normal software execution and for exceptions. The
currently active impure data pointer is stored in an extra variable.
During the exception handler the data pointer is replaced temporarily
and later restored.

Furthermore this enables dynamic reentrancy, meaning there is a
function __getreent() that returns the currently active impure data
pointer. This is not really necessary, but flexible with respect to
the future multicore implementation.