or1k: Add malloc lock to newlib
As malloc may be called during software runtime or during an
exception, it is necessary to lock the access. In single core
environments it is sufficient to disable all interrupts during the
execution of malloc. This patch does this. The implementation also
keeps track of the current lock holder and allows recursion. Actually
both the exception and the software will never try to lock the malloc
while it is locked by the other one, as they are locked out during the
entire duration of the first call to __malloc_lock until the end of
the last call to __malloc_unlock. But this is the generic
implementation that eases extension to multicore.